+++ src/main/java/OGTSystem/entity/MessageEntity.java
@@ -1,9 +1,6 @@
 package OGTSystem.entity;
 
 import lombok.Data;
-import org.springframework.web.bind.annotation.RequestParam;
-
-import java.math.BigInteger;
 
 @Data
 public class MessageEntity {
@@ -13,7 +10,7 @@ public class MessageEntity {
     // 消息目标ID
     private String uuidto;
     // 消息目标NO
-    private BigInteger uunoto;
+    private Long uunoto;
     // 来源用户token
     private String token;
     // 消息类型
