+++ src/main/java/OGTSystem/entity/UserAuthEntity.java
@@ -2,7 +2,6 @@ package OGTSystem.entity;
 
 import lombok.Data;
 
-import java.math.BigInteger;
 
 @Data
 public class UserAuthEntity {
@@ -10,7 +9,7 @@ public class UserAuthEntity {
     // 唯一用户ID
     private String UUID;
     // 自增唯一用户编号
-    private BigInteger UUNO;
+    private Long UUNO;
     // 用户token
     private String userToken;
     // 用户违规次数
